Move the stresser logic into dedicated file
